server/pyproject.toml
[tool.poetry]
name = "hackerthon"
version = "0.1.0"
description = ""
authors = ["MINMO <ku0629@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
selenium = "^4.27.1"
django-cors-headers = "^4.6.0"
transformers = "^4.46.3"
datasets = "^3.1.0"
torch = "^2.5.1"
accelerate = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
